requirements.txt
streamlit
numpy
pandas
Pillow
tensorflow
opencv-python